[libretiny] Move HAL bodies into components/libretiny/hal.cpp + inline trivial dispatches
Per-platform follow-up to #15977.

Move out-of-line HAL bodies (delayMicroseconds, arch_init, arch_restart)
from components/libretiny/core.cpp into a new components/libretiny/hal.cpp.
core.cpp is now empty (no extra component bootstrap to keep) — left as a
stub for symmetry.

Inline the trivial one-liners directly in hal_libretiny.h with forward
decls for the LibreTiny C API:

  arch_feed_wdt()              -> lt_wdt_feed()
  arch_get_cpu_cycle_count()   -> lt_cpu_get_cycle_count()
  arch_get_cpu_freq_hz()       -> lt_cpu_get_freq()

arch_init stays out-of-line because of the libretiny::setup_preferences()
call (component-private header) plus the BK72xx priority-raise + GPIO
recovery logic.

delayMicroseconds() also stays out-of-line — Arduino's ::delayMicroseconds
signature is unsigned int and forward-declaring it across the C/C++
boundary in a hot header is fragile.
