+++ esphome/core/hal/hal_libretiny.h
@@ -53,6 +53,13 @@ extern "C" unsigned long micros(void);
 extern "C" unsigned long millis(void);
 // NOLINTEND(google-runtime-int,readability-identifier-naming,readability-redundant-declaration)
 
+// Forward decls from libretiny's <lt_api.h> family for the inline arch_*
+// wrappers below. Pulling the full header would drag in the rest of the
+// LibreTiny C API.
+extern "C" void lt_wdt_feed(void);
+extern "C" uint32_t lt_cpu_get_cycle_count(void);
+extern "C" uint32_t lt_cpu_get_freq(void);
+
 namespace esphome {
 
 /// Returns true when executing inside an interrupt handler.
@@ -88,11 +95,12 @@ __attribute__((always_inline)) inline uint32_t millis() { return static_cast<uin
 #endif
 __attribute__((always_inline)) inline uint64_t millis_64() { return Millis64Impl::compute(millis()); }
 
+__attribute__((always_inline)) inline void arch_feed_wdt() { lt_wdt_feed(); }
+__attribute__((always_inline)) inline uint32_t arch_get_cpu_cycle_count() { return lt_cpu_get_cycle_count(); }
+__attribute__((always_inline)) inline uint32_t arch_get_cpu_freq_hz() { return lt_cpu_get_freq(); }
+
 void delayMicroseconds(uint32_t us);  // NOLINT(readability-identifier-naming)
-void arch_feed_wdt();
-uint32_t arch_get_cpu_cycle_count();
 void arch_init();
-uint32_t arch_get_cpu_freq_hz();
 
 }  // namespace esphome
